# docs/api/openapi3.yaml
openapi: 3.0.1
info:
  title: sync
  description: ""
  version: 0.0.1
servers:
- url: ""
...
paths:
...
  /posts:
    get:
      tags:
      - post
      summary: Get Posts
      description: Get Posts
      operationId: GetPosts
      parameters:
      - name: first
        in: query
        description: Number of items to return from the beginning of the list
        required: false
        schema:
          type: string
      - name: after
        in: query
        description: "Cursor for pagination, indicating the position after which to\
          \ return items"
        required: false
        schema:
          type: string
      - name: last
        in: query
        description: Number of items to return from the end of the list
        required: false
        schema:
          type: string
      - name: before
        in: query
        description: "Cursor for pagination, indicating the position before which\
          \ to return items"
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPostsResponse"
              examples:
                GetPosts:
                  value: "{\"posts\":{\"pageInfo\":{\"size\":1,\"hasNextPage\":false,\"\
                    hasPreviousPage\":false,\"startCursor\":\"start-cursor\",\"endCursor\"\
                    :\"end-cursor\"},\"nodes\":[{\"cursor\":\"cursor\",\"content\"\
                    :{\"id\":1,\"slug\":\"test-slug\",\"type\":\"SHORT\",\"scope\"\
                    :\"PUBLIC\",\"status\":\"PUBLISHED\",\"title\":\"title\",\"author\"\
                    :{\"name\":\"User\",\"handle\":\"user-handle\"},\"project\":{\"\
                    handle\":\"project-handle\",\"name\":\"Project\"},\"content\"\
                    :\"Post Content\",\"resolved\":false,\"createdAt\":\"1970-01-01T00:00:00Z\"\
                    }}]}}"
    post:
      tags:
      - post
      summary: Create Post
      description: Create Post
      operationId: CreatePost
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
            examples:
              CreatePost:
                value: "{\"title\":\"title\",\"type\":\"SHORT\",\"scope\":\"WORKSPACE\"\
                  ,\"status\":\"PUBLISHED\",\"content\":{\"text\":\"This is a post\
                  \ content.\",\"json\":\"{\\\"text\\\": \\\"This is a post content.\\\
                  \"}\",\"mediaIds\":[1,2]},\"tags\":[\"java\",\"spring\"],\"project\"\
                  :{\"handle\":\"project-handle\"}}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePostResponse"
              examples:
                CreatePost:
                  value: "{\r\n  \"slug\" : \"slug\"\r\n}"
  /posts/{postId}:
    delete:
      tags:
      - post
      summary: Delete Post
      description: Delete Post
      operationId: DeletePost
      parameters:
      - name: postId
        in: path
        description: Post ID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
    patch:
      tags:
      - post
      summary: Update Post
      description: Update Post
      operationId: UpdatePost
      parameters:
      - name: postId
        in: path
        description: Post ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostRequest"
            examples:
              UpdatePost:
                value: "{\"content\":\"This is a post content.\"}"
      responses:
        "204":
          description: "204"
  /posts/{slug}:
    get:
      tags:
      - post
      summary: Get Post By Slug
      description: Get Post By Slug
      operationId: GetPostBySlug
      parameters:
      - name: slug
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPostResponse"
              examples:
                GetPostBySlug:
                  value: "{\"id\":1,\"type\":\"SHORT\",\"scope\":\"PUBLIC\",\"status\"\
                    :\"PUBLISHED\",\"slug\":\"test-slug\",\"title\":\"title\",\"author\"\
                    :{\"name\":\"Author Name\",\"handle\":\"author-handle\"},\"project\"\
                    :{\"handle\":\"project-handle\",\"name\":\"Project Name\"},\"\
                    content\":{\"json\":\"Post Content\",\"media\":[{\"id\":1,\"url\"\
                    :\"https://example.com/media.png\"}]},\"likeCount\":1,\"commentCount\"\
                    :1,\"bookmarked\":true,\"resolved\":false,\"createdAt\":\"1970-01-01T00:00:00Z\"\
                    }"
...
  /projects/{handle}/posts:
    get:
      tags:
      - post
      summary: Get Posts By Project
      description: Get Posts By Project
      operationId: GetPostsByProject
      parameters:
      - name: handle
        in: path
        description: 프로젝트 핸들
        required: true
        schema:
          type: string
      - name: first
        in: query
        description: Number of items to return from the beginning of the list
        required: false
        schema:
          type: string
      - name: after
        in: query
        description: "Cursor for pagination, indicating the position after which to\
          \ return items"
        required: false
        schema:
          type: string
      - name: last
        in: query
        description: Number of items to return from the end of the list
        required: false
        schema:
          type: string
      - name: before
        in: query
        description: "Cursor for pagination, indicating the position before which\
          \ to return items"
        required: false
        schema:
          type: string
      - name: type
        in: query
        description: 게시글 타입
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPostsResponse"
              examples:
                GetPostsByProject:
                  value: "{\"posts\":{\"pageInfo\":{\"size\":1,\"hasNextPage\":false,\"\
                    hasPreviousPage\":false,\"startCursor\":\"start-cursor\",\"endCursor\"\
                    :\"end-cursor\"},\"nodes\":[{\"cursor\":\"cursor\",\"content\"\
                    :{\"id\":1,\"slug\":\"test-slug\",\"type\":\"SHORT\",\"scope\"\
                    :\"PUBLIC\",\"status\":\"PUBLISHED\",\"title\":\"title\",\"author\"\
                    :{\"name\":\"User\",\"handle\":\"user-handle\"},\"project\":{\"\
                    handle\":\"project-handle\",\"name\":\"Project\"},\"content\"\
                    :\"Post Content\",\"resolved\":false,\"createdAt\":\"1970-01-01T00:00:00Z\"\
                    }}]}}"
...
  /users/{userId}/posts:
    get:
      tags:
      - post
      summary: Get Posts
      description: Get Posts
      operationId: GetUserPosts
      parameters:
      - name: userId
        in: path
        description: User ID
        required: true
        schema:
          type: string
      - name: first
        in: query
        description: Number of items to return from the beginning of the list
        required: false
        schema:
          type: string
      - name: after
        in: query
        description: "Cursor for pagination, indicating the position after which to\
          \ return items"
        required: false
        schema:
          type: string
      - name: last
        in: query
        description: Number of items to return from the end of the list
        required: false
        schema:
          type: string
      - name: before
        in: query
        description: "Cursor for pagination, indicating the position before which\
          \ to return items"
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPostsResponse"
              examples:
                GetUserPosts:
                  value: "{\"posts\":{\"pageInfo\":{\"size\":1,\"hasNextPage\":false,\"\
                    hasPreviousPage\":false,\"startCursor\":\"start-cursor\",\"endCursor\"\
                    :\"end-cursor\"},\"nodes\":[{\"cursor\":\"cursor\",\"content\"\
                    :{\"id\":1,\"slug\":\"test-slug\",\"type\":\"SHORT\",\"scope\"\
                    :\"PUBLIC\",\"status\":\"PUBLISHED\",\"title\":\"title\",\"author\"\
                    :{\"name\":\"User\",\"handle\":\"user-handle\"},\"project\":{\"\
                    handle\":\"project-handle\",\"name\":\"Project\"},\"content\"\
                    :\"Post Content\",\"resolved\":false,\"createdAt\":\"1970-01-01T00:00:00Z\"\
                    }}]}}"
components:
  schemas:
...
    GetPostsResponse:
      title: GetPostsResponse
      type: object
      properties:
        posts:
          required:
          - nodes
          - pageInfo
          type: object
          properties:
            nodes:
              type: array
              description: Nodes
              items:
                required:
                - content
                - cursor
                type: object
                properties:
                  cursor:
                    type: string
                    description: Node Cursor
                  content:
                    required:
                    - author
                    - content
                    - createdAt
                    - id
                    - resolved
                    - scope
                    - slug
                    - status
                    - type
                    type: object
                    properties:
                      createdAt:
                        type: string
                        description: Creation Timestamp
                      author:
                        required:
                        - handle
                        - name
                        type: object
                        properties:
                          name:
                            type: string
                            description: Author Name
                          handle:
                            type: string
                            description: Author Handle
                        description: Author Information
                      scope:
                        type: string
                        description: Post Scope
                        enum:
                        - PUBLIC
                        - WORKSPACE
                      project:
                        required:
                        - handle
                        - name
                        type: object
                        properties:
                          name:
                            type: string
                            description: Project Name
                          handle:
                            type: string
                            description: Project Handle
                        description: Project Information
                      id:
                        type: number
                        description: Post ID
                      title:
                        type: string
                        description: Post Title
                        nullable: true
                      type:
                        type: string
                        description: Post Type
                        enum:
                        - SHORT
                        - LONG
                        - QUESTION
                      content:
                        type: string
                        description: Post Content
                      slug:
                        type: string
                        description: Post Slug
                      resolved:
                        type: boolean
                        description: Whether the question post has been resolved
                      status:
                        type: string
                        description: Post Status
                        enum:
                        - DRAFT
                        - PUBLISHED
                    description: Node Content
            pageInfo:
              required:
              - hasNextPage
              - hasPreviousPage
              - size
              type: object
              properties:
                size:
                  type: number
                  description: Page Size
                hasNextPage:
                  type: boolean
                  description: Has Next Page
                hasPreviousPage:
                  type: boolean
                  description: Has Previous Page
                endCursor:
                  type: string
                  description: End Cursor
                  nullable: true
                startCursor:
                  type: string
                  description: Start Cursor
                  nullable: true
              description: Page Info
...
    CreatePostResponse:
      title: CreatePostResponse
      required:
      - slug
      type: object
      properties:
        slug:
          type: string
          description: Post Slug
...
    CreatePostRequest:
      title: CreatePostRequest
      required:
      - content
      - type
      type: object
      properties:
        scope:
          type: string
          description: Post Scope
          nullable: true
          enum:
          - PUBLIC
          - WORKSPACE
        project:
          type: object
          properties:
            handle:
              type: string
              description: 프로젝트 핸들
              nullable: true
          description: 프로젝트
        type:
          type: string
          description: Post Type
          enum:
          - SHORT
          - LONG
          - QUESTION
        title:
          type: string
          description: Title
          nullable: true
        content:
          required:
          - json
          - text
          type: object
          properties:
            json:
              type: string
              description: JSON Content
            mediaIds:
              type: array
              description: 사용된 미디어 ID 목록
              nullable: true
              items:
                oneOf:
                - type: object
                - type: boolean
                - type: string
                - type: number
            text:
              type: string
              description: Text Content
          description: Content
        tags:
          type: array
          description: 태그 목록
          nullable: true
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        status:
          type: string
          description: Post Status
          nullable: true
          enum:
          - DRAFT
          - PUBLISHED
...
    GetPostResponse:
      title: GetPostResponse
      required:
      - author
      - bookmarked
      - commentCount
      - content
      - createdAt
      - id
      - likeCount
      - resolved
      - scope
      - slug
      - status
      - type
      type: object
      properties:
        bookmarked:
          type: boolean
          description: Whether the current user bookmarked this post
        author:
          required:
          - handle
          - name
          type: object
          properties:
            name:
              type: string
              description: Author Name
            handle:
              type: string
              description: Author Handle
          description: Author Information
        project:
          required:
          - handle
          - name
          type: object
          properties:
            name:
              type: string
              description: Project Name
            handle:
              type: string
              description: Project Handle
          description: Project Information
        likeCount:
          type: number
          description: Number of Likes
        title:
          type: string
          description: Post Title
          nullable: true
        type:
          type: string
          description: Post Type
          enum:
          - SHORT
          - LONG
          - QUESTION
        content:
          required:
          - json
          - media
          type: object
          properties:
            json:
              type: string
              description: Post Content JSON
            media:
              type: array
              description: Media attached to the post
              items:
                required:
                - id
                - url
                type: object
                properties:
                  id:
                    type: number
                    description: Media ID
                  url:
                    type: string
                    description: Media URL
          description: Post Content
        commentCount:
          type: number
          description: Number of Comments
        createdAt:
          type: string
          description: Creation Timestamp
        scope:
          type: string
          description: Post Scope
          enum:
          - PUBLIC
          - WORKSPACE
        id:
          type: number
          description: Post ID
        slug:
          type: string
          description: Post Slug
        resolved:
          type: boolean
          description: Whether the question post has been resolved
        status:
          type: string
          description: Post Status
          enum:
          - DRAFT
          - PUBLISHED
...
    UpdatePostRequest:
      title: UpdatePostRequest
      required:
      - content
      type: object
      properties:
        content:
          type: string
          description: Content
...